Update postupdate.sh
Adds the ability to stop /storage/joypads from being cleared out, if a file dont_delete_me is specifically created there, hence it should be fine if there still really is a reason for that to be in postupdate.sh

+++ packages/351elec/sources/scripts/postupdate.sh
@@ -116,8 +116,9 @@ fi
 
 ## 2021-09-30:
 ## Remove any configurd ES joypads on upgrade
-rm -f /storage/joypads/*.cfg
-
+if [ ! -f /storage/joypads/dont_delete_me ]; then
+  rm -f /storage/joypads/*.cfg
+fi
 
 ## 2021-09-27:
 ## Force replacement of gamecontrollerdb.txt on update
